changed int32 to int8 for memory saving

+++ Go_Forza_Rec/source/parser/parser.go
@@ -10,7 +10,7 @@ import (
 )
 
 type recievedcarstate struct {
-	IsRaceOn         int32
+	IsRaceOn         int8
 	TimestampMS      uint32
 	EngineMaxRPM     float32
 	EngineIdleRPM    float32
